+++ include/IECoreDelight/ShaderNetworkAlgo.h
@@ -48,6 +48,12 @@ namespace ShaderNetworkAlgo
 
 IECOREDELIGHT_API IECoreScene::ShaderNetworkPtr preprocessedNetwork( const IECoreScene::ShaderNetwork *shaderNetwork );
 
+/// Returns the NSI type of the geometry needed for the light or `nullptr` if none is needed.
+const char *lightGeometryType( const IECoreScene::ShaderNetwork *shaderNetwork );
+/// Edits the light geometry, which will have already been created with the right type.
+/// `state` is used to store the current state and avoid unnecessary edits.
+void updateLightGeometry( const IECoreScene::ShaderNetwork *shaderNetwork, NSIContext_t context, const char *handle, IECore::MurmurHash &state );
+
 }  // namespace ShaderNetworkAlgo
 
 }  // namespace IECoreDelight